refactor writing test ouput based on feedback
