+++ 06-building-a-program-start-to-finish/writing-html-js-css-outputs-(template)/output/index.html
@@ -16,7 +16,7 @@
 
 </head>
 
-<body ng-controller="IndexController">
+<body ng-controller="IndexCtrl">
     <div class="image-container set-full-height" style="background-image: url('lib/images/banner.jpg')">
         <div class="content white">
             <nav class="navbar navbar-default" role="navigation">
@@ -98,7 +98,7 @@
                 </div>
             </div>
             <div class="col-sm-8 col-sm-offset-0">
-                <div class="card" ng-show="welcomePage">
+                <div class="card" ng-show="welcome">
                     <h4 class="main-title">Welcome, Let's Get Started</h4>
                     <center>
                         <div class="welcome-text">Inverted Index is designed to allow very fast full-text searches. An Inverted Index consists of a
@@ -155,7 +155,7 @@
                             <tr ng-repeat="token in tokens">
                                 <td style="text-transform: capitalize;">{{token}}</td>
                                 <td ng-repeat="book in bookIndices">
-                                    <img ng-src="/lib/images/{{ book | imageCheck : indices[token] }}.png" width="20px" height="20px">
+                                    <img ng-src="/lib/images/{{ book | rowFunc : indices[token] }}.png" width="20px" height="20px">
                                 </td>
                             </tr>
                         </tbody>
@@ -182,7 +182,7 @@
                             <tr ng-repeat="(keyword, searchValue) in fileValue.result">
                                 <td style="text-transform: capitalize;">{{keyword}}</td>
                                 <td ng-repeat="val in fileValue.countBook">
-                                    <img ng-src="/lib/images/{{ val | imageCheck : searchValue}}.png" width="20px" height="20px">
+                                    <img ng-src="/lib/images/{{ val | rowFunc : searchValue}}.png" width="20px" height="20px">
                                 </td>
                             </tr>
                         </tbody>
@@ -196,7 +196,7 @@
        <center><p style="font-size: 1.2em; color: #0f0f0f">≈Måximuƒ≈</p></center>
        <footer class="main-footer">
       <div class="pull-right hidden-xs">
-        <b>Version</b> 1.1.5
+        <b>Version</b> 1.0.0
       </div>
       <strong>Copyright &copy; 2017 <a href="http://github.com/andela-ksolomon">Solomon Kingsley</a>.</strong> All rights reserved.
     </footer>
@@ -209,8 +209,7 @@
 <script type="text/javascript" src="lib/js/bootstrap.min.js"></script>
 <script type="text/javascript" src="lib/js/angular.min.js"></script>
 <script type="text/javascript" src="src/js/app.js"></script>
-<script type="text/javascript" src="/lib/js/main.js"></script>
-<script type="text/javascript" src="src/js/invertedIndex.js"></script>
+<script type="text/javascript" src="lib/js/main.js"></script>
+<script type="text/javascript" src="src/js/InvertedIndex.js"></script>
 <script type="text/javascript" src="/lib/js/dir-pagination.js"></script>
-
 </html>
